requirements.txt
requests==2.23.0
pytest==5.0.1
allure-pytest==2.8.12
allure-python-commons==2.8.12
pytest-dependency==0.5.1
